Added fallback server option
As the main server seems to be having some issues, I've added an option to use a fallback server when the main one fails.

This defaults to "" (undefined) in the config, with users being able to set the URL below the standard URL option in options.

If you've got testingServer set, this is overridden, and the form is removed from the options.

How it works:

Upon requesting sponsors, if the main server fails, we call retryFetch to try again. As we keep note of how many times we retry, we use the fallback server on every even retry, so it flips back and forth.

The delay between retries has also been adjusted. 404 is the same, but if we've failed 2 or less times, we retry instantly, before going back to the previous delay setting.

As most users are only concerned about fetching sponsors, we don't use the fallback server for requests regarding submitting data.

We also only use the fallback server on asyncRequestToServer() calls, as nothing using the synchronous method would benefit.